refactor: Модуль bodyload перенесен в папку additions
Сделано по аналогии с файлами стилей — чтобы отделить изначальный
код от изменений при выполнении заданий курса
