+++ source/js/script.js
@@ -9,9 +9,6 @@ import form from './modules/form.js';
 import social from './modules/social.js';
 import FullPageScroll from './modules/full-page-scroll';
 
-// custom modules
-import bodyload from './modules/bodyload.js';
-
 // init modules
 mobileHeight();
 slider();
@@ -21,7 +18,11 @@ chat();
 result();
 form();
 social();
-bodyload();
 
 const fullPageScroll = new FullPageScroll();
 fullPageScroll.init();
+
+// custom modules
+import bodyload from './additions/bodyload.js';
+
+bodyload();
